Created entity-show with minimum structure
